refactor: decouple validation schemas from default settings to tree shake
